Fix rendering of the details
